--- .travis.yml
@@ -1,35 +1,0 @@
-language: python
-python:
-  - '3.8'
-
-os:
-  - linux
-env:
-  global:
-    - TX_BRANCH="master"
-install:
-  - pip install poetry
-  - pip install transifex-client sphinx
-  - pip install coveralls codecov
-  - poetry install -v
-
-script:
-  - poetry run pytest --cov="src" -s tests
-
-after_success:
-  - coveralls --rcfile=.coverage
-  - codecov
-  - cd ..
-  - cd docs
-  - if [[ $TRAVIS_PULL_REQUEST == "false" ]]; then sh .tx/tx.sh push; sh .tx/tx.sh pull;
-    fi
-  - cd ..
-deploy:
-  provider: pypi
-  user: julien.malard
-  password:
-    secure: pXM8GUCauL2GI1r2cinSSSEDk9/Golsz94W5pYv+NuzfPgWMOjRJq2UqQPNk+ElHRn0c69muW8AZ7jPTR+XP9CpXn4deYv8aQ0TCJ5aRutmFx1lTGXSHSqH2RN6LOsurAs8LI1V4kp3nYXUno4WxMVWPzr7O0exd+ao5dJkmHYAYZ06k4Uaflr9qFtJKawYuQoejVAWBGVHWr7d4GnEnIqXl3Wzz6gFBWv9Oeftubwt8fzN6XW9OEHkzE5xfooOdpKjKwWpSvO9pJu0W1OheupUlLSnuncATVPGmr4cEbsUl3UHrPeiSmdSUa3kEXHluFsdSoXHnSh6O0T1ta3oYatuOWIP+le8sW67erWZ8qmtZhA7dverVzaKFmAFhJ/8epxftI0Geeth3jtEc/X+aURCPlE/3RGbJIk+jEz1CnZ7zAz0tAUs/T4Ct9bA11+F7ikDNU2EOy8sO98/DBOeYRnDJXWU31EA2CMI21uLJJM+FxxiC1xRnH7TNZO3ZQz7JTjlnkV45VnmPSpMbXT54khbXchaDChph++6CNE53jCAHbPazel1Q1fOBtkXWPCogmFDplzGg++papCV8AtlaPQc4C75JkKZfOtllcWF40upW0OBci26vLT1TCISe7mhG9/Df0YGjOhUxrO7VdCLWhFaCB+RytsGN7KQFtnGhJBU=
-  distributions: sdist bdist_wheel
-  skip_existing: true
-  on:
-    tags: true
