--- appveyor.yml
@@ -1,28 +1,0 @@
-environment:
-
-  matrix:
-    - PYTHON: "C:\\Python37-x64"
-    - PYTHON: "C:\\Python37"
-
-install:
-  - "set PATH=%PATH%;%PYTHON%\\Scripts"
-  - "%PYTHON%\\python.exe -m pip install --upgrade pip"
-  - "%PYTHON%\\python.exe -m pip install --upgrade setuptools"
-  - "%PYTHON%\\python.exe -m pip install --upgrade wheel"
-  - "git clone https://github.com/julienmalard/ennikkai.git"
-  - ps: cd ennikkai
-  - "%PYTHON%\\python.exe setup.py -q install"
-  - ps: cd ..
-  - "git clone https://github.com/julienmalard/taqdir.git"
-  - ps: cd taqdir
-  - "%PYTHON%\\python.exe setup.py -q install"
-  - ps: cd ..
-  - "%PYTHON%\\python.exe -m pip install -r reqs.txt"
-  - "%PYTHON%\\python.exe -m pip install -r reqs.op.txt"
-  - "%PYTHON%\\python.exe setup.py -q install"
-
-build: false
-
-test_script:
-  - ps: cd pruebas
-  - "%PYTHON%\\python.exe -m unittest"
