+++ pyproject.toml
@@ -11,16 +11,16 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.8"
-trio = "^0.17.0"
-click = "^7.1.2"
+python = ">=3.10,<3.12"
+trio = "0.22.0"
+click = "^8.1.3"
 numpy = "^1.19.2"
-xarray = "^0.16.1"
+xarray = "2023.1.0"
 ennikkai = "^1.2.5"
-pysd = "^0.10.0"
+pysd = "^3.9.0"
 pandas = "^1.1.2"
-scipy = "^1.5.2"
-regex = "^2020.7.14"
+scipy = "^1.10.0"
+regex = "^2022.10.31"
 python-dateutil = "^2.8.1"
 asks = "^2.4.10"
 tldextract = "^2.2.3"
